#114 Sublayers Object, togglable uncertainty ellipses, configurable rover waypoint iamges
