+++ src/essence/Tools/Layers/LayersTool.css
@@ -182,7 +182,6 @@
     margin: 0;
 }
 #layersTool .settings ul > li {
-    height: 24px;
     line-height: 24px;
 }
 #layersTool .settings ul > li > div {
